manual fix after merge
